default repo root to the symlink base in certain cases
when the link is to a subdirectory of the git repo, the contraction
failed. This change fixes that by taking the symlink base as the repo
root (since the real repo root can't be accessed).

e.g.:
Take `~/.config/fish -> ~/projects/dotfiles/shell/fish/`. If `dotfiles`
is the true git root, `fish` would be recognized as the repo root since
any further ancestors would be outside the symlink path.